Extract store interface to its own file
