+++ packages/xstate-wallet/src/store/memory-store.ts
@@ -22,6 +22,7 @@ import {Chain, FakeChain} from '../chain';
 import {calculateChannelId, hashState} from './state-utils';
 import {NETWORK_ID} from '../constants';
 import {checkThat, exists} from '../utils';
+import {Store} from './store';
 
 interface DirectFunding {
   type: 'Direct';
@@ -60,36 +61,6 @@ interface InternalEvents {
   channelUpdated: [ChannelStoreEntry];
   newObjective: [Objective];
   addToOutbox: [Message];
-}
-
-export interface Store {
-  newObjectiveFeed: Observable<Objective>;
-  outboxFeed: Observable<Message>;
-  pushMessage: (message: Message) => Promise<void>;
-  channelUpdatedFeed(channelId: string): Observable<ChannelStoreEntry>;
-
-  getAddress(): string;
-  signAndAddState(channelId: string, stateVars: StateVariables): Promise<void>;
-  createChannel(
-    participants: Participant[],
-    challengeDuration: BigNumber,
-    stateVars: StateVariables,
-    appDefinition?: string
-  ): Promise<ChannelStoreEntry>;
-  getEntry(channelId): Promise<ChannelStoreEntry>;
-  getLedger(peerId: string): Promise<ChannelStoreEntry>;
-  // TODO: This is awkward. Might be better to set the funding on create/initialize channel?
-  setFunding(channelId: string, funding: Funding): Promise<void>;
-
-  // TODO: I don't know how the store is mean to send outgoing messages.
-  // But I need one, in order to implement virtual funding.
-  addObjective(objective: Objective): void;
-
-  // TODO: should this be exposed via the Store?
-  chain: Chain;
-
-  getBudget: (site: string) => Promise<SiteBudget | undefined>;
-  updateOrCreateBudget: (budget: SiteBudget) => Promise<void>;
 }
 
 export class MemoryStore implements Store {
