+++ packages/xstate-wallet/src/workflows/application.ts
@@ -17,7 +17,7 @@ import {sendDisplayMessage, MessagingServiceInterface, convertToChannelResult} f
 import {filter, map, tap, flatMap, first} from 'rxjs/operators';
 import * as CCC from './confirm-create-channel';
 import {createMockGuard, getDataAndInvoke} from '../utils/workflow-utils';
-import {Store} from '../store/memory-store';
+import {Store} from '../store';
 import {StateVariables} from '../store/types';
 import {ChannelStoreEntry} from '../store/channel-store-entry';
 import {bigNumberify} from 'ethers/utils';
