+++ packages/xstate-wallet/src/workflows/tests/application.test.ts
@@ -3,7 +3,8 @@ import {ethers} from 'ethers';
 import waitForExpect from 'wait-for-expect';
 import {applicationWorkflow, WorkflowServices, WorkflowActions} from '../application';
 import {AddressZero} from 'ethers/constants';
-import {MemoryStore, Store} from '../../store/memory-store';
+import {MemoryStore} from '../../store/memory-store';
+import {Store} from '../../store';
 import {StateVariables, SignedState} from '../../store/types';
 import {ChannelStoreEntry} from '../../store/channel-store-entry';
 import {MessagingService, MessagingServiceInterface} from '../../messaging';
